LayoutTests/imported/w3c/web-platform-tests/css/css-tables/col_removal-expected.txt

x	x

FAIL Table grid syncs after COL removal assert_equals: table shrinks after column is removed expected 64 but got 137
